move implementations + apiDump
